requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
requests==2.33.1
openai>=1.0.0
apscheduler>=3.10.4
